Rotation properties of Spherical Harmonics
A new file has been added. This file defines
objects which allow the computation of how spherical
harmonics behave under point group operations.

The implementation is "naive" and uses a formula
which is known to be unstable for large angular momentum l.
However, I cannot foresee using l>3, so it should be fine
for now. Other, more stable algorithms are available, but
seem much more difficult to implement.